# EasyTrufi.Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backend EasyTrufi API",
    "description": "Documentación de la API de EasyTrufi - .NET 8",
    "contact": {
      "name": "Equipo de Desarrollo UCB",
      "email": "carlos.camacho.h@ucb.edu.bo"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Payment": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "Recupera una lista paginada de pagos realizados en el sistema.",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "NfcCardId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "UserId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "AmountCents",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "ValidatorId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna la lista de pagos",
            "schema": {
              "$ref": "#/definitions/PaymentDTOIEnumerableApiResponse"
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "Inserta un nuevo pago en el sistema.",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "El objeto DTO que contiene los datos del pago a insertar.",
            "schema": {
              "$ref": "#/definitions/PaymentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pago creado exitosamente",
            "schema": {
              "$ref": "#/definitions/PaymentApiResponse"
            }
          },
          "400": {
            "description": "Datos inválidos para la creación del pago",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Payment/{id}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "Recupera un pago específico por su identificador único.",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "El identificador único del pago.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna el pago solicitado",
            "schema": {
              "$ref": "#/definitions/PaymentDTOApiResponse"
            }
          },
          "404": {
            "description": "Pago no encontrado",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Payment"
        ],
        "summary": "Actualiza los datos de un pago existente.",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "El identificador único del pago a actualizar.",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "El objeto DTO que contiene los nuevos datos del pago.",
            "schema": {
              "$ref": "#/definitions/PaymentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pago actualizado exitosamente",
            "schema": {
              "$ref": "#/definitions/PaymentApiResponse"
            }
          },
          "404": {
            "description": "Pago no encontrado",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Payment"
        ],
        "summary": "Elimina un pago del sistema.",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "El identificador único del pago a eliminar.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Pago eliminado exitosamente"
          },
          "404": {
            "description": "Pago no encontrado",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Message": {
      "type": "object",
      "properties": {
        "type": {
          "description": "Tipo del mensaje (e.g., Información, Advertencia, Error)",
          "type": "string"
        },
        "description": {
          "description": "Descripción detallada del mensaje",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "NfcCard": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "uid": {
          "type": "string"
        },
        "userId": {
          "format": "int64",
          "type": "integer"
        },
        "active": {
          "type": "boolean"
        },
        "issuedAt": {
          "format": "date-time",
          "type": "string"
        },
        "payments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Payment"
          }
        },
        "topups": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Topup"
          }
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    },
...
    "Pagination": {
      "type": "object",
      "properties": {
        "totalCount": {
          "format": "int32",
          "description": "Número total de elementos en la colección",
          "type": "integer"
        },
        "pageSize": {
          "format": "int32",
          "description": "Cantidad de elementos por página",
          "type": "integer"
        },
        "currentPage": {
          "format": "int32",
          "description": "Número de la página actual",
          "type": "integer"
        },
        "totalPages": {
          "format": "int32",
          "description": "Número total de páginas",
          "type": "integer"
        },
        "hasNextPage": {
          "description": "Indica si existe una página siguiente",
          "type": "boolean"
        },
        "hasPreviousPage": {
          "description": "Indica si existe una página anterior",
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "Payment": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "nfcCardId": {
          "format": "int64",
          "type": "integer"
        },
        "userId": {
          "format": "int64",
          "type": "integer"
        },
        "amountCents": {
          "format": "int64",
          "type": "integer"
        },
        "validatorId": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "nfcCard": {
          "$ref": "#/definitions/NfcCard"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "validator": {
          "$ref": "#/definitions/Validator"
        }
      },
      "additionalProperties": false
    },
    "PaymentApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Payment"
        },
        "pagination": {
          "$ref": "#/definitions/Pagination"
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        }
      },
      "additionalProperties": false
    },
    "PaymentDTO": {
      "type": "object",
      "properties": {
        "nfcCardId": {
          "format": "int64",
          "type": "integer"
        },
        "userId": {
          "format": "int64",
          "type": "integer"
        },
        "amountCents": {
          "format": "int64",
          "type": "integer"
        },
        "validatorId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "PaymentDTOApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/PaymentDTO"
        },
        "pagination": {
          "$ref": "#/definitions/Pagination"
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        }
      },
      "additionalProperties": false
    },
    "PaymentDTOIEnumerableApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PaymentDTO"
          }
        },
        "pagination": {
          "$ref": "#/definitions/Pagination"
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        }
      },
      "additionalProperties": false
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "Topup": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "nfcCardId": {
          "format": "int64",
          "type": "integer"
        },
        "userId": {
          "format": "int64",
          "type": "integer"
        },
        "amountCents": {
          "format": "int64",
          "type": "integer"
        },
        "method": {
          "type": "string"
        },
        "reference": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "completedAt": {
          "format": "date-time",
          "type": "string"
        },
        "nfcCard": {
          "$ref": "#/definitions/NfcCard"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "cedula": {
          "description": "Número de Cédula de Identidad (C.I.) del usuario",
          "type": "string"
        },
        "fullName": {
          "description": "Nombre completo del usuario (Nombres y Apellidos)",
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date",
          "description": "Fecha de nacimiento (usado para validar edad o descuentos)",
          "type": "string"
        },
        "email": {
          "description": "Dirección de correo electrónico del usuario",
          "type": "string"
        },
        "passwordHash": {
          "description": "Hash encriptado de la contraseña del usuario",
          "type": "string"
        },
        "emailVerified": {
          "description": "Indica si el usuario ha verificado su correo electrónico",
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "description": "Fecha de registro del usuario en el sistema",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "description": "Fecha de la última modificación de los datos del usuario",
          "type": "string"
        },
        "nfcCards": {
          "description": "Lista de tarjetas NFC vinculadas a este usuario",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NfcCard"
          }
        },
        "payments": {
          "description": "Historial de pagos de pasajes realizados",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Payment"
          }
        },
        "topups": {
          "description": "Historial de recargas de crédito en la cuenta",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Topup"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Validator": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "validatorCode": {
          "description": "Código único de identificación del dispositivo validador",
          "type": "string"
        },
        "vehicleId": {
          "description": "Placa o ID del trufi/miniván asociado",
          "type": "string"
        },
        "locationDescription": {
          "description": "Descripción operativa (Ej: 'Línea 210 - Trufi 45')",
          "type": "string"
        },
        "active": {
          "description": "Indica si el validador está habilitado para procesar cobros",
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "description": "Fecha de registro e instalación del validador",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "description": "Fecha de última actualización de datos del dispositivo",
          "type": "string"
        },
        "payments": {
          "description": "Colección de transacciones realizadas en este validador",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Payment"
          }
        }
      },
      "additionalProperties": false
    },
...
  }
}